requirements.txt
fastapi
uvicorn
pandas
numpy
sentence-transformers
faiss-cpu
openpyxl
scikit-learn